Fix Qt6 build issues in Qt models and payment server

+++ src/qt/assettablemodel.cpp
@@ -208,7 +208,7 @@ QVariant AssetTableModel::data(const QModelIndex &index, int role) const
         case Qt::TextAlignmentRole:
         {
             if (index.column() == Quantity) {
-                return Qt::AlignRight | Qt::AlignVCenter;
+                return static_cast<int>(Qt::AlignRight | Qt::AlignVCenter);
             }
         }
         default:
@@ -231,9 +231,9 @@ QVariant AssetTableModel::headerData(int section, Qt::Orientation orientation, i
             return QSize(30, 50);
     } else if (role == Qt::TextAlignmentRole) {
         if (orientation == Qt::Vertical)
-            return Qt::AlignLeft | Qt::AlignVCenter;
+            return static_cast<int>(Qt::AlignLeft | Qt::AlignVCenter);
 
-        return Qt::AlignHCenter | Qt::AlignVCenter;
+        return static_cast<int>(Qt::AlignHCenter | Qt::AlignVCenter);
     }
 
     return QVariant();
